requirements.txt
Django<1.8
django-bootstrap3
django-mptt
django-static-precompiler
djangorestframework
markdown
django-filter
